src/requirements.txt
adbc-driver-manager==0.6.0
adbc-driver-sqlite==0.6.0
altgraph==0.17.3
annotated-types==0.5.0
asttokens==2.4.0
backcall==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
connectorx==0.3.2
contourpy==1.1.0
cx-Freeze==6.15.7
cx-Logging==3.1.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
deltalake==0.10.2
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.42.1
fsspec==2023.9.0
greenlet==2.0.2
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lief==0.13.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
Nuitka==1.8.1
numpy==1.25.2
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
polars==0.19.2
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pyinstaller @ file:///C:/Users/d00858458/Downloads/pyinstaller-5.13.2/pyinstaller-develop/pyinstaller-develop
pyinstaller-hooks-contrib==2023.8
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.1
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
xlsx2csv==0.8.1
XlsxWriter==3.1.3
zstandard==0.21.0
